Moved PartyContactDetails & ShipmentLocation to model package.
