+++ src/main/java/org/dcsa/core/events/repository/PartyContactDetailsRepository.java
@@ -1,6 +1,6 @@
 package org.dcsa.core.events.repository;
 
-import org.dcsa.core.events.PartyContactDetails;
+import org.dcsa.core.events.model.PartyContactDetails;
 import org.dcsa.core.repository.ExtendedRepository;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
